pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
]
